move schedule tests in a common folder
